Renamed Navbar to more approriate Header. Added logo, tweaks to navigation, Button component and others
